renaming basefactors to attention_multiple
